Threat_Intel_Project/requirements.txt
requests 
aiohttp 
pandas 
elasticsearch 
python-dotenv 
rich 
schedule
whois
ipwhois
python-whois